se arreglo el bug de doble renderizado en el /favorite y se creo un spinner de carga, a su vez se hizo una completa refactorizacion del codigo para las peliculas favoritas
